requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.14.38
botocore==1.17.38
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
dataclasses-json==0.5.2
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
en-core-web-sm==2.3.1
entrypoints==0.3
enum-compat==0.0.3
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.12
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
matplotlib==3.3.0
mistune==0.8.4
murmurhash==1.0.2
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.1.1
numpy==1.19.1
oauth2client==3.0.0
oauthlib==3.1.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.14.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
ptyprocess==0.6.0
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-youtube==0.6.4
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
smart-open==2.1.0
smmap==3.0.4
spacy==2.3.2
srsly==1.0.2
streamlit==0.74.1
stringcase==1.2.0
terminado==0.8.3
testpath==0.4.4
texthero==1.0.9
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
validators==0.17.1
wasabi==0.7.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
